requirements.txt
pyinstaller
python-dotenv
jinja2
build twine
upgrade setuptools wheel twine
